apps/ai-service/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector Search (RAG)
pgvector>=0.2.4

# AI/Agent
pydantic-ai>=0.0.10
google-generativeai>=0.3.2

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0

# Rate Limiting
slowapi>=0.1.9

# File Upload
aiofiles>=23.0.0
pypdf>=4.0.0
python-multipart>=0.0.6

# Utilities
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.1

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
aiosqlite>=0.19.0
ruff>=0.1.14
mypy>=1.8.0
